backend/core/apps/modules/serializers.py
from rest_framework import serializers
from .models import Module

class ModuleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Module
        fields = ['id', 'title', 'dictionary','is_public']
        
        
        
