backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0 
django-environ==0.12.0     
django-filter==25.1        
django-rest==0.8.7
djangorestframework==3.15.2
Markdown==3.7
psycopg2==2.9.10
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
gevent==24.11.1
gunicorn==23.0.0
django-debug-toolbar==5.1.0
redis==5.2.1
hiredis==3.1.0
django-redis==5.4.0